refactor(prompts): Abstract MCP error parsing into separate module
All existing functionality preserved with clean module interface.

🤖 Assisted by Amazon Q Developer
